misc: Add a file to ignore revisions on git blame
Add the file .git-blame-ignore-revs to store the hash
of the commits that should be ignored on a git blame.
The commit hash should be added after the commit has
been merged. New hashes should be added as the first
line of the file.

Add some initial contents with commits that only fix
the style.

Usage example:

  git blame file.py --ignore-revs-file .git-blame-ignore-revs

To always ignore from this file on git blames:

  git config blame.ignoreRevsFile .git-blame-ignore-revs

Change-Id: Idd29cb933d4157f1650658acc7efc1bd142fa7c7
Signed-off-by: Daniel R. Carvalho <odanrc@yahoo.com.br>
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/43594
Reviewed-by: Jason Lowe-Power <power.jg@gmail.com>
Maintainer: Jason Lowe-Power <power.jg@gmail.com>
Tested-by: kokoro <noreply+kokoro@google.com>